Improve structure and documentation of Hits connector
