+++ Sources/InstantSearchCore/Hits/Connector/HitsConnector+GeoSearch.swift
